requirements.txt
python-telegram-bot==20.7
ping3==4.0.4
mysql-connector-python==8.1.0
python-dotenv==1.0.0
cryptography==41.0.7
httpx==0.25.2
schedule==1.2.1